requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.6
frozenlist==1.5.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-tools==1.69.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.5
langsmith==0.2.10
marshmallow==3.25.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.1
openai==1.60.0
orjson==3.10.14
packaging==24.2
pluggy==1.5.0
portalocker==2.10.1
propcache==0.2.1
protobuf==5.29.3
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyPDF2==3.0.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-dotenv==1.0.1
python-multipart==0.0.20
pywin32==308
PyYAML==6.0.2
qdrant-client==1.12.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
